Create EncyptDrive.java for LevelOneEncrypt
This is the driver/main entry point of the program. It takes in user input and displays the result of Bitshifter. This program was built only as a demo and not for any real use, nor should it ever be used for any sort of real encryption, it is far too simple.